Broke menus into their own component.
This is working for layout-main.  Next will be to modify the other layout managers to remove their menu logic and just use the shared component.
