+++ CSETWebNg/src/app/layout/layoutmain/layout-main.component.scss
@@ -1370,7 +1370,7 @@ ul.nav-tabs li.active button .tab-icon {
   color: $gray-150;
   background-color: transparent;
   border: none;
-  text-decoration: none;
+  text-decoration: none !important;
   height: 2.2rem;
   line-height: 2.2rem;
 }
@@ -1381,7 +1381,7 @@ ul.nav-tabs li.active button .tab-icon {
 .menu-panel [mat-menu-item].mat-menu-item-highlighted:not([disabled]) {
   color: $primary-color;
   background-color: $white;
-  text-decoration: none !important;
+  // text-decoration: none !important;
 }
 
 .menu-panel [mat-menu-item].mat-menu-item.disabled-menu-item,
@@ -1389,6 +1389,7 @@ ul.nav-tabs li.active button .tab-icon {
 .menu-panel [mat-menu-item].mat-menu-item.disabled-menu-item:hover {
   color: $primary-250;
   background-color: transparent;
+  // text-decoration: none !important;
   cursor: not-allowed !important;
 }
