+++ CSETWebNg/src/app/layout/layoutmain/layout-main.component.ts
@@ -23,7 +23,6 @@
 ////////////////////////////////
 import { Component, OnInit, ViewChild, AfterViewInit, ViewEncapsulation } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
-import { MatMenuModule } from '@angular/material/menu';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Hotkey, HotkeysService } from 'angular2-hotkeys';
 import { AboutComponent } from '../../dialogs/about/about.component';
@@ -71,35 +70,15 @@ export class LayoutMainComponent implements OnInit, AfterViewInit {
     public aggregationSvc: AggregationService,
     public fileSvc: FileUploadClientService,
     public dialog: MatDialog,
-    public router: Router,
-    private _hotkeysService: HotkeysService
+    public router: Router
   ) { }
 
-
-  ngOnInit() {
-    this.docUrl = this.configSvc.docUrl;
-
-    if (localStorage.getItem("returnPath")) {
-      if (!Number(localStorage.getItem("redirectid"))) {
-        this.hasPath(localStorage.getItem("returnPath"));
-      }
-    }
-
-    this.setupShortCutKeys();
-  }
-
+  ngOnInit() { }
+  
   ngAfterViewInit() {
-
     setTimeout(() => {
       this.isFooterOpen();
     }, 200);
-  }
-
-  hasPath(rpath: string) {
-    if (rpath != null) {
-      localStorage.removeItem("returnPath");
-      this.router.navigate([rpath], { queryParamsHandling: "preserve" });
-    }
   }
 
   /**
@@ -124,240 +103,18 @@ export class LayoutMainComponent implements OnInit, AfterViewInit {
     return false;
   }
 
-
   goHome() {
     this.assessSvc.dropAssessment();
     this.router.navigate(['/home']);
-  }
-
-  /**
-   * Show the RRA tutorial in a dialog.  This is temporary, until
-   * a proper User Guide is written for RRA.
-   * @returns 
-   */
-  ransomwareReadiness() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    this.dialogRef = this.dialog.open(RraMiniUserGuideComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  /**
-   * 
-   * @returns 
-   */
-  about() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    this.dialogRef = this.dialog.open(AboutComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  /**
-   * 
-   * @returns 
-   */
-  termsOfUse() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    this.dialogRef = this.dialog.open(TermsOfUseComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  advisory() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    this.dialogRef = this.dialog.open(AdvisoryComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  editUser() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    this.dialogRef = this.dialog.open(EditUserComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe(
-        (data: CreateUser) => {
-          // the update user request happened when the dialog's form was saved
-          this.dialogRef = undefined;
-        },
-        error => console.log(error.message)
-      );
-  }
-
-  resetPassword() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    this.dialogRef = this.dialog.open(ChangePasswordComponent, {
-      width: '300px',
-      data: { primaryEmail: this.auth.email() }
-    });
-    this.dialogRef.afterClosed().subscribe();
   }
 
   inAssessment() {
     return localStorage.getItem('assessmentId');
   }
 
-  showAssessDocs() {
-    if (this.dialog.openDialogs[0]) {
-      return;
-    }
-    // , {width: '800px', height: "500px"}
-    this.dialogRef = this.dialog.open(AssessmentDocumentsComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  editParameters() {
-    if (this.dialog.openDialogs[0]) {
-
-      return;
-    }
-    this.dialogRef = this.dialog.open(GlobalParametersComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  enableProtectedFeature() {
-    if (this.dialog.openDialogs[0]) {
-
-      return;
-    }
-    this.dialogRef = this.dialog.open(EnableProtectedComponent);
-  }
-
-  showKeyboardShortcuts() {
-    if (this.dialog.openDialogs[0]) {
-
-      return;
-    }
-    this.dialogRef = this.dialog.open(KeyboardShortcutsComponent);
-  }
-
-  showExcelExportDialog() {
-    const doNotShowLocal = localStorage.getItem('doNotShowExcelExport');
-    const doNotShow = doNotShowLocal && doNotShowLocal == 'true' ? true : false;
-    if (this.dialog.openDialogs[0] || doNotShow) {
-      this.exportToExcel();
-      return;
-    }
-    this.dialogRef = this.dialog.open(ExcelExportComponent);
-    this.dialogRef
-      .afterClosed()
-      .subscribe();
-  }
-
-  exportToExcel() {
-    window.location.href = this.configSvc.apiUrl + 'ExcelExport?token=' + localStorage.getItem('userToken');
-  }
-
-  navigateTrend() {
-    this.aggregationSvc.mode = 'TREND';
-    this.router.navigate(['/trend']);
-  }
-
-  navigateCompare() {
-    this.aggregationSvc.mode = 'COMPARE';
-    this.router.navigate(['/compare']);
-  }
 
   // -----------------------------
-
-  setupShortCutKeys() {
-    // About
-    this._hotkeysService.add(new Hotkey('alt+a', (event: KeyboardEvent): boolean => {
-      this.about();
-      return false; // Prevent bubbling
-    }));
-    // Accessibility Features
-    this._hotkeysService.add(new Hotkey('alt+c', (event: KeyboardEvent): boolean => {
-      window.open(this.docUrl + "ApplicationDocuments/AccessibilityStatement.pdf", "_blank");
-      return false; // Prevent bubbling
-    }));
-    // User Guide
-    this._hotkeysService.add(new Hotkey('alt+g', (event: KeyboardEvent): boolean => {
-      window.open(this.docUrl + "htmlhelp/index.htm", "_blank");
-      return false; // Prevent bubbling
-    }));
-    // Resource Library
-    this._hotkeysService.add(new Hotkey('alt+l', (event: KeyboardEvent): boolean => {
-      this.router.navigate(['resource-library']);
-      return false; // Prevent bubbling
-    }));
-    // Parameter Editor
-    this._hotkeysService.add(new Hotkey('alt+m', (event: KeyboardEvent): boolean => {
-      return false; // Prevent bubbling
-    }));
-    // New Assessment
-    this._hotkeysService.add(new Hotkey('alt+n', (event: KeyboardEvent): boolean => {
-      const dialogRef = this.dialog.open(ConfirmComponent);
-      dialogRef.componentInstance.confirmMessage =
-        "Are you sure you want to create a new assessment? ";
-      dialogRef.afterClosed().subscribe(result => {
-        if (result) {
-          this.assessSvc.newAssessment();
-        }
-      });
-      return false; // Prevent bubbling
-    }));
-    // User Guide (PDF)
-    this._hotkeysService.add(new Hotkey('alt+p', (event: KeyboardEvent): boolean => {
-      window.open(this.docUrl + "cdDocs/UserGuide.pdf", "_blank");
-      return false; // Prevent bubbling
-    }));
-
-    // Questionnaires
-    // this._hotkeysService.add(new Hotkey('alt+q', (event: KeyboardEvent): boolean => {
-    //   return false; // Prevent bubbling
-    // }));
-
-    // Protected Features
-    this._hotkeysService.add(new Hotkey('alt+r', (event: KeyboardEvent): boolean => {
-      this.enableProtectedFeature();
-      return false; // Prevent bubbling
-    }));
-    // Assessment Documents
-    this._hotkeysService.add(new Hotkey('alt+t', (event: KeyboardEvent): boolean => {
-      return false; // Prevent bubbling
-    }));
-    // Advisory
-    this._hotkeysService.add(new Hotkey('alt+v', (event: KeyboardEvent): boolean => {
-      this.advisory();
-      return false; // Prevent bubbling
-    }));
-    // Keyboard Shortcuts
-    this._hotkeysService.add(new Hotkey('?', (event: KeyboardEvent): boolean => {
-      this.showKeyboardShortcuts();
-      return false; // Prevent bubbling
-    }));
-
-  }
-
-  showNavBarRight() {
-    if (this.auth.isLocal) {
-      return false;
-    }
-
-    return this.router.url !== '/resource-library';
-  }
+ 
 
   isFooterOpen() {
     if (!!this.accordion) {
